Updated behavior of the Create and Change Password Forms
Updated the functionality of the create post and change password forms to be more like the update post form. where after a successful submission the form will dim to prevent additional submits. The dimmed feature will clear the next time the form is pulled to add a new post or change password